test-requirements.txt
click>=7.0
pyinstaller>=3.5
ipython>=7.8
pyyaml==5.1.2